# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EpicResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "linearProjectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linearprojectid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "repositoryId": {
            "title": "Repositoryid",
            "type": "integer"
          },
          "rootBranch": {
            "title": "Rootbranch",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repositoryId",
          "name",
          "slug",
          "rootBranch",
          "linearProjectId",
          "createdAt"
        ],
        "title": "EpicResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HarnessProfileUpsertRequest": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/HarnessProfileDefinitionResponse"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/HarnessProfileSource",
            "default": "user"
          }
        },
        "required": [
          "id",
          "kind",
          "displayName",
          "definition"
        ],
        "title": "HarnessProfileUpsertRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Redesmyn",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/epics/{epic}": {
      "get": {
        "operationId": "get_epic_v1_epics__epic__get",
        "parameters": [
          {
            "in": "path",
            "name": "epic",
            "required": true,
            "schema": {
              "title": "Epic",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpicResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Epic"
      }
    },
    "/v1/epics/{epic}/graph": {
      "get": {
        "operationId": "epic_graph_v1_epics__epic__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "epic",
            "required": true,
            "schema": {
              "title": "Epic",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpicGraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Epic Graph"
      }
    },
    "/v1/harness-profiles": {
      "get": {
        "operationId": "list_harness_profiles_v1_harness_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HarnessProfileResponse"
                  },
                  "title": "Response List Harness Profiles V1 Harness Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Harness Profiles"
      },
      "post": {
        "operationId": "upsert_harness_profile_v1_harness_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessProfileUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Harness Profile"
      }
    },
...
    "/v1/hosts/upsert": {
      "post": {
        "operationId": "upsert_host_v1_hosts_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Host"
      }
    },
...
    "/v1/nodes/{node_id}/agent": {
      "post": {
        "operationId": "set_node_agent_v1_nodes__node_id__agent_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeSetAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Node Agent"
      }
    },
...
    "/v1/tasks/{task_id}/agent/restart": {
      "post": {
        "operationId": "restart_task_agent_v1_tasks__task_id__agent_restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAgentRestartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAgentStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart Task Agent"
      }
    },
    "/v1/tasks/{task_id}/agent/start": {
      "post": {
        "operationId": "start_task_agent_v1_tasks__task_id__agent_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAgentStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAgentStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Task Agent"
      }
    },
    "/v1/tasks/{task_id}/agent/stop": {
      "post": {
        "operationId": "stop_task_agent_v1_tasks__task_id__agent_stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAgentStopResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Task Agent"
      }
    }
  }
}